requirements.txt
# ============================================================
# requirements.txt
# Streamlit Cloud가 앱을 배포할 때 자동으로 설치할 라이브러리 목록입니다.
# 각 줄에 라이브러리 이름을 하나씩 적으면 됩니다.
# ============================================================

# 웹 화면(UI)을 만들어주는 핵심 라이브러리
streamlit

# Google Gemini AI API를 사용하기 위한 라이브러리
google-generativeai
